+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 3.4.0
+
+* add Freeze Method to ObjectFactory
+* add ForSegment Method to Array
+
 ### 3.3.0
 
 * add Difference Function in ObjectFactory